20180720: try to test

+++ kvraft/common.go
@@ -26,7 +26,7 @@ type PutAppendReply struct {
 type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
-	Nonce int32
+	Nonce int64
 }
 
 type GetReply struct {
